refact(discord): move config options from client to configManager
