+++ package.json
@@ -25,6 +25,7 @@
     "@types/compression": "^1.7.5",
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
+    "@types/json-bigint": "^1.0.4",
     "@types/jsonwebtoken": "^9.0.2",
     "@types/morgan": "^1.9.4",
     "@types/node": "^20.4.8",
@@ -43,6 +44,7 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-rate-limit": "^6.8.1",
+    "json-bigint": "^1.0.0",
     "jsonwebtoken": "^9.0.1",
     "mariadb": "^3.2.0",
     "morgan": "^1.10.0",
